This is my portfolio website.
This static website is composed of 5 sections:
- Home
- Skills
- Projects
- About
- Testaments
- Contact

Already completed the design for the Home section, Skills section, Contact sections.

The layout on the Projects section is in progress.

On the Contact section, there are not yet any functionalities as of the moment.

This version mainly focuses as of the moment for the 1441px display and beyond.

The colors that were used on the website are #ebf5ee, #283044, #064789 and #427aa1.